Refactor: 알림 메시지 Resolver 구조를 Map 기반으로 변경
